Fix SoF runner type check, integers and nulls
Trickle mode skips resources whose type differs from the view's
resource. Integer cells hold 64-bit values so integer64 does not
overflow. SQLite storage writes SQL NULL for empty integer and boolean
cells and keeps date/dateTime values in FHIR string form. Unsupported
FHIRPath host services raise FHIRException instead of java.lang.Error.

+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/utils/sql/StorageJson.java
@@ -55,7 +55,7 @@ public class StorageJson implements Storage {
     if (value == null) {
       return new JsonNull();
     } else if (value.getValueInt() != null) {
-      return new JsonNumber(value.getValueInt().intValue());
+      return new JsonNumber(value.getValueInt().longValue());
     }
     if (value.getValueBoolean() != null) {
       return new JsonBoolean(value.getValueBoolean().booleanValue());
